Rename DesignSystem page -> DesignSystemGuide
